refactor: Replace Function with Function
originally the put http method was being used however as per convention this is incorrect. PATCH is to be used when not replacing an entire object, within Django's code partial_update does call update in any case simply with the partial flag to include not a complete replacement as it is unnecessary.