test: removed dead code / test
the test in question calls the api using the put http request. This functionality was removed however the call itself is incorrect as either put or patch require an id to be specified. A sanity test is already included the current test file and if the test is changed to behave properly as a standard api call would be made (patch or put with the id) the request goes through thus it should be removed as the purpose of this test is no longer needed.