feat: added new View for Custom User
This adds flexibility to now call the API from the route to filter whatever results desired, include whichever fields you want returned and whatever fields you want saved or excluded as well. All the information sent must exist within the database before hand. This will help reduce the creation of multiple serializers for simple changes in fields desired. This is meant as a replacement for the CustomUser