revert: original serializes back
The reason for the revert although the previous code was working is a django limitation was encountered. In order to have a many to may relationship, the models must be declared within the Class and are not available to set at run time / dynamically. If this is attempted, the many to many relationship is lost and only the linking foreign key is returned.