feat: added security via permissions
this moves the api calls into their respective viewsets to match the permissions that have been setup. Because of this, the url has also been changed to keep the api calls under the audit/ url. One test was affected by the url change and the mobile client has been updated to retrieve the information appropriately