refactor: extract variable from local storage
To avoid using strings to get the values from the local storage. This helps us keep track of what values will be in the local storage and allows us to use the compiler to help us write out the locations in the storage instead of remembering them and writing them out by hand which is error prone.