refactor: move System Language field
It didn't make sense to be passing in the system language to each individual component, since the language will most likely be handled by a single class for all the components. Therefore, I decided to remove the constructor parameter and have the language be decided and used in the abstract language factory.